Änderung float und Gleitkommazahlen, zzgl. neue Projekt  .Datentypen_Privat

+++ MeinRechner/Program.cs
@@ -17,12 +17,12 @@ namespace MeinRechner
          Console.Write("Bitte gib den zweiten Summanden ein: ");
          string zweiterSummand = Console.ReadLine();
 
-         //  Wandel Text in Ganzzahlen
-         int ersterSummandAlsZahl = Convert.ToInt32(ersterSummand);
-         int zweiterSummandAlsZahl = Convert.ToInt32(zweiterSummand);
+         //  Wandel Text in Gleitkommazahlen
+         float ersterSummandAlsZahl = Convert.ToSingle(ersterSummand);
+         float zweiterSummandAlsZahl = Convert.ToSingle(zweiterSummand);
 
          // Berechnung ausführen
-         int summe = ersterSummandAlsZahl + zweiterSummandAlsZahl;
+         float summe = ersterSummandAlsZahl + zweiterSummandAlsZahl;
 
          //Ausgabe
          Console.WriteLine("Die summe beträgt: {0}", summe);
